pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geminiwebcli"
version = "0.1.0"
description = "CLI tool for Gemini via browser automation"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "prompt_toolkit>=3.0",
    "rich>=13.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
geminiwebcli = "geminiwebcli.cli:main"

[project.optional-dependencies]
bot = ["python-telegram-bot>=20.0"]

[tool.setuptools.packages.find]
where = ["."]
